Refactor NightWriter class out of night_writer runner.
