+++ spec/spec_helper.rb
@@ -1,0 +1,4 @@
+require 'simplecov'
+SimpleCov.start
+
+require './spec/night_writer_spec'
